refactor: split biome and biome-mapper loaders into separate modules
