+++ src/game/assets/worldgen/mod.rs
@@ -1,0 +1,5 @@
+pub mod biome;
+pub mod biome_mapper;
+
+pub use biome::*;
+pub use biome_mapper::*;
